requirements.txt
netmiko>=4.1.2
textfsm>=1.1.3
napalm>=4.1.0
pyyaml>=6.0.1
sqlalchemy>=2.0.0
python-dotenv>=1.0.0 